Update
The target="_blank" attribute was the issue that was causing the links to not change colour when visited, so I removed it from all of them.

There is a fix for this that still opens links in a new tab, which involves JavaScript, but I am not learning that right now.